+++ toolbox/speakeasy2.m
@@ -1,15 +1,15 @@
-function [memb, ordering] = speakeasy2(g, opts)
+function [memb, ordering] = speakeasy2(graph, graphOpts, opts, attribute)
 %SPEAKEASY2 community detection
-%   MEMBERSHIP = SPEAKEASY2(G) Cluster the graph G.
+%   MEMBERSHIP = SPEAKEASY2(GRAPH) Cluster the GRAPH.
 %
-%   [MEMB, ORDERING] = SPEAKEASY2(G) provide an index vector to order nodes
+%   [MEMB, ORDERING] = SPEAKEASY2(GRAPH) provide an index vector to order nodes
 %   based on the community they are in, in order from largest community to
 %   smallest. This can be used with HEATMAP or IMAGESC to help display
-%   communities.
+%   communities. Ordering can also be created with SE2.ORDER.
 %
-%   SPEAKEASY2('VERSION') display the toolbox version.
-%
-%   VERSION = SPEAKEASY2('VERSION') return the toolbox version string.
+%   GRAPH = SPEAKEASY2(GRAPH, ..., 'result', NAME) if GRAPH can hold an
+%   attribute (if represented by a MATLAB graph, default), the results are
+%   attached to the graph as node attribute NAME.
 %
 %   MEMBERSHIP = SPEAKEASY2(G, 'PARAM1', VAL1, 'PARAM2', VAL2, ...) the
 %   behavior of speakeasy2 can be modified through name-value pairs. Available
@@ -52,8 +52,8 @@ function [memb, ordering] = speakeasy2(g, opts)
 %   See also SE2.ORDER
 
     arguments
-        g {mustBeAdjOrVersion};
-        opts.isdirected (1, 1) logical = isdirected(g);
+        graph {igutils.mustBeGraph};
+        graphOpts.?igutils.GraphInProps;
         opts.discardTransient (1, 1) {mustBePositive, mustBeInteger} = 3;
         opts.independentRuns (1, 1) {mustBePositive, mustBeInteger} = 10;
         opts.maxThreads (1, 1) {mustBeNonnegative, mustBeInteger} = 0;
@@ -63,19 +63,15 @@ function [memb, ordering] = speakeasy2(g, opts)
         opts.subcluster (1, 1) {mustBePositive, mustBeInteger} = 1;
         opts.minCluster (1, 1) {mustBePositive, mustBeInteger} = 5;
         opts.targetClusters (1, 1) {mustBePositive, mustBeInteger} = ...
-            defaultTargetClusters(g);
+            defaultTargetClusters(graph);
         opts.targetPartitions (1, 1) {mustBePositive, mustBeInteger} = 5;
         opts.verbose = false;
+        attribute.results (1, :) char ...
+            {igutils.mustHoldNodeAttr(graph, attribute.results)} = '';
     end
 
-    if ischar(g) || isstring(g)
-        if nargout == 0
-            se2.version();
-        else
-            memb = se2.version();
-        end
-        return
-    end
+    graphOpts = namedargs2cell(graphOpts);
+    graphOpts = igutils.setGraphInProps(graph, graphOpts{:});
 
     if opts.multiCommunity ~= 1
         error("SE2:NotImplemented", ...
@@ -88,14 +84,19 @@ function [memb, ordering] = speakeasy2(g, opts)
     end
 
     if nargout == 2
-        [memb, ordering] = mexSE2(g, opts);
+        [memb, ordering] = mexSE2(graph, opts, graphOpts);
     else
-        memb = mexSE2(g, opts);
+        memb = mexSE2(graph, opts, graphOpts);
+    end
+
+    if ~isempty(attribute.results)
+        graph.Nodes.(attribute.results) = memb';
+        memb = graph;
     end
 end
 
-function n = defaultTargetClusters(g)
-    nNodes = length(g);
+function n = defaultTargetClusters(graph)
+    nNodes = igraph.numnodes(graph);
     n = ceil(nNodes / 100);
     if n < 10
         n = 10;